+++ plugins/sudoers/toke_util.c
@@ -172,7 +172,7 @@ fill_args(const char *s, int len, int addspace)
     p = sudoerslval.command.args + arg_len;
     if (addspace)
 	*p++ = ' ';
-    if (strlcpy(p, s, arg_size - (p - sudoerslval.command.args)) != len) {
+    if (strlcpy(p, s, arg_size - (p - sudoerslval.command.args)) != (size_t)len) {
 	warningx(_("fill_args: buffer overflow"));	/* paranoia */
 	sudoerserror(NULL);
 	debug_return_bool(false);
